% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummulative_function.R
\name{cummax_date}
\alias{cummax_date}
\title{Cummulative maximum}
\usage{
cummax_date(x = structure(c(NA, NA, NA, NA, 17126, 18351, 16839, 17621, 18337,
  NA), class = "Date"))
}
\arguments{
\item{x}{vector of date object}
}
\description{
This function returns a vector whose elements are the maxima of the elements of the argument for date object
}
\examples{
cummax_date()
}
\keyword{cummax_date}

