% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummulative_function.R
\name{cummin_date}
\alias{cummin_date}
\title{Cummulative minimum}
\usage{
cummin_date(x = structure(c(NA, NA, NA, NA, 17126, 18351, 16839, 17621, 18337,
  NA), class = "Date"))
}
\arguments{
\item{x}{vector of date object}
}
\description{
This function returns a vector whose elements are the minima of the elements of the arguments for date object
}
\examples{
cummin_date()
}
\keyword{cummin_date}

